refactor[#235]: 핸들러 객체 위치 변경
