+++ src/main/java/JJinBBang/app/global/security/handler/CustomAccessDeniedHandler.java
@@ -1,4 +1,4 @@
-package JJinBBang.app.global.security.exception;
+package JJinBBang.app.global.security.handler;
 
 import java.io.IOException;
 
@@ -7,6 +7,8 @@ import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.web.access.AccessDeniedHandler;
 import org.springframework.stereotype.Component;
 
+import JJinBBang.app.global.security.exception.SecurityAccessDeniedException;
+import JJinBBang.app.global.security.exception.SecurityErrorResponder;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
